requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
dj-database-url==2.2.0
Django==5.1
django-bs4==0.0.4
django-ckeditor==6.7.1
django-filter==24.3
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-taggit==6.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
pywebio==1.8.3
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
user-agents==2.2.0
